Move 3 main pages into pages folder
