+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react"
 import "./styles/App.css"
-import Home from "./components/Home"
+import Home from "./pages/Home"
 
 function App() {
   return (
