+++ src/pages/Home.js
@@ -1,4 +1,4 @@
-import NavBar from "./NavBar"
+import NavBar from "../components/NavBar"
 
 function Home() {
   return (
